move sourcecode out of the public folder
